+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- Added `CharSequence.codePointSequence(startIndex: Int, endIndex: Int)`
+
 ## [0.8.0] - 2024-06-09
 ### Changed
 - Updated to Kotlin 2.0.0